backend/requirements.txt
teler==0.2.0
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.8.0
eventlet==0.33.3
python-dotenv==1.0.0
requests==2.31.0